+++ confik/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add support for `try_from` attribute, following the rules of `from` but using `TryFrom`. Note: This required a change in the return type of `try_build` and a new `Error` variant.
+  - This will not break existing code unless it contains manual implementations of `Configuration`.
+
 ## 0.10.2
 
 - Remove `Debug` implementation from configuration builders.